refactor: simplify injector
Injector does not care whether something is provided as a class or
function. It just calls `provider.create(args)` and it's
a responsibility of the concrete provider to instantiate it as a class
or function.

Therefore the whole "super constructor business" is not responsibility
of injector but rather `ClassProvider`. It normalizes parameters of
all the constructors (so that the injector does not care).