chore(package.json): use traceur#0.0.33 until dependencies can cope with 0.0.66
There are a number of breaking changes in more modern traceur versions, such as:

- sourceMap option renamed to sourceMaps
- filename option removed, it is now a parameter of the compile function
- result.errors may not be defined / not an array if no errors occur
- for some reason, either tests are not being preprocessed, or else karma-traceur-preprocessor
  freaks out when it sees the "import" keyword in tests. (I am not sure yet if this is a traceur
  issue, or a bug in pipe or something).

Rather than wait for fixes in each of the dependency projects affected by this break, instead
we will just use a LKGR of traceur, which seems to work well enough. Fixes in the dependency
tools should get this working, but perhaps not until a traceur#v0.1.0 ships.

Closes #70
Closes #71