fix: check prototype first to conclude if its a class
After minifying your javascript it is possible that the class names get changed, such that they are no more adhering to the standard of starting a class name with a capital letter. Thus the isClass util fun returns false in such cases, even though the class has some fields set in the prototype object.

The solution is to first check the prototype object for keys and then check for class name.

Fixes #92